Extracted all animations to a dedicated package, moved reusable modifiers to Components.kt
