+++ app/src/main/java/com/developersbreach/composeactors/ui/components/DynamicThemeGenerator.kt
@@ -1,16 +1,11 @@
-package com.developersbreach.composeactors.utils
+package com.developersbreach.composeactors.ui.components
 
 import android.content.Context
-import androidx.annotation.FloatRange
 import androidx.compose.animation.animateColorAsState
 import androidx.compose.animation.core.Spring
 import androidx.compose.animation.core.spring
 import androidx.compose.material.MaterialTheme
 import androidx.compose.runtime.*
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.composed
-import androidx.compose.ui.draw.drawBehind
-import androidx.compose.ui.graphics.Brush
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.graphics.compositeOver
 import androidx.compose.ui.graphics.luminance
@@ -25,62 +20,11 @@ import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.withContext
 import kotlin.math.max
 import kotlin.math.min
-import kotlin.math.pow
 
 /**
  * This file belongs to official jetpack samples repository made by google.
  * Sample -> JetCaster.
  */
-
-
-/**
- * Draws a vertical gradient scrim in the foreground.
- *
- * @param color The color of the gradient scrim.
- * @param startYPercentage The start y value, in percentage of the layout's height (0f to 1f)
- * @param endYPercentage The end y value, in percentage of the layout's height (0f to 1f)
- * @param decay The exponential decay to apply to the gradient. Defaults to `1.0f` which is
- * a linear gradient.
- * @param numStops The number of color stops to draw in the gradient. Higher numbers result in
- * the higher visual quality at the cost of draw performance. Defaults to `16`.
- */
-fun Modifier.verticalGradientScrim(
-    color: Color,
-    @FloatRange(from = 0.0, to = 1.0) startYPercentage: Float = 0f,
-    @FloatRange(from = 0.0, to = 1.0) endYPercentage: Float = 1f,
-    decay: Float = 1.0f,
-    numStops: Int = 16
-): Modifier = composed {
-    val colors = remember(color, numStops) {
-        if (decay != 1f) {
-            // If we have a non-linear decay, we need to create the color gradient steps
-            // manually
-            val baseAlpha = color.alpha
-            List(numStops) { i ->
-                val x = i * 1f / (numStops - 1)
-                val opacity = x.pow(decay)
-                color.copy(alpha = baseAlpha * opacity)
-            }
-        } else {
-            // If we have a linear decay, we just create a simple list of start + end colors
-            listOf(color.copy(alpha = 0f), color)
-        }
-    }
-
-    var height by remember { mutableStateOf(0f) }
-    val brush = remember(color, numStops, startYPercentage, endYPercentage, height) {
-        Brush.verticalGradient(
-            colors = colors,
-            startY = height * startYPercentage,
-            endY = height * endYPercentage
-        )
-    }
-
-    drawBehind {
-        height = size.height
-        drawRect(brush = brush)
-    }
-}
 
 @Composable
 private fun rememberDominantColorState(
@@ -103,7 +47,7 @@ private fun rememberDominantColorState(
  * @param isColorValid A lambda which allows filtering of the calculated image colors.
  */
 @Stable
-class DominantColorState(
+private class DominantColorState(
     private val context: Context,
     private val defaultColor: Color,
     private val defaultOnColor: Color,
@@ -202,7 +146,7 @@ private suspend fun calculateSwatchesInImage(
 }
 
 @Composable
-fun ActorDynamicTheme(
+fun ImageBackgroundThemeGenerator(
     podcastImageUrl: String,
     content: @Composable () -> Unit
 ) {
