+++ app/src/main/java/com/developersbreach/composeactors/ui/components/Progress.kt
@@ -9,9 +9,9 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
-import com.developersbreach.composeactors.ui.search.AnimatedSearch
-import com.developersbreach.composeactors.utils.ApiKey
-import com.developersbreach.composeactors.utils.InfinitelyFlowingCircles
+import com.developersbreach.composeactors.animations.AnimatedSearch
+import com.developersbreach.composeactors.utils.TmdbApiKey
+import com.developersbreach.composeactors.animations.InfinitelyFlowingCircles
 
 /**
  * @param isLoadingData if true circular progress bar will show.
@@ -21,7 +21,7 @@ import com.developersbreach.composeactors.utils.InfinitelyFlowingCircles
 fun ShowProgressIndicator(
     isLoadingData: Boolean
 ) {
-    if (isLoadingData && ApiKey.API_KEY.isNotEmpty()) {
+    if (isLoadingData && TmdbApiKey.TMDB_API_KEY.isNotEmpty()) {
         Box(
             modifier = Modifier.fillMaxSize()
         ) {
